+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Post v1 a change is further considered breaking if the state file format is inco
 * Moved messages to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
 * Moved role names, descriptions and rules to external files using QuasiQuotes. ([#155](https://github.com/hjwylde/werewolf/issues/155))
 * Removed `--include-seer` option from `start`. ([#223](https://github.com/hjwylde/werewolf/issues/223))
+* Removed `--force` option from `end`. ([#224](https://github.com/hjwylde/werewolf/issues/224))
 
 ### v1.1.1.0